refactor: move scripts into backend package
