+++ backend/__init__.py
@@ -1,0 +1,25 @@
+"""Backend package for reusable workflow functions."""
+
+from .yolo import run_prediction, run_training
+from .collect_images import collect_images, setup_logger as setup_collect_logger
+from .convert_yolo_to_ls import convert_yolo_to_ls
+from .download_annotated import download_annotated
+from .generate_data_yaml import generate_data_yaml
+from .index_predictions_by_class import index_predictions
+from .stage_predictions_for_upload import stage_predictions
+from .train_yolo import run_training as run_full_training
+from .upload_gcs import upload as upload_gcs
+
+__all__ = [
+    "run_prediction",
+    "run_training",
+    "collect_images",
+    "setup_collect_logger",
+    "convert_yolo_to_ls",
+    "download_annotated",
+    "generate_data_yaml",
+    "index_predictions",
+    "stage_predictions",
+    "run_full_training",
+    "upload_gcs",
+]
